tomcat: readd sourcing of setenv.sh in init script for v5.5 package
We don't replace the catalina.sh (dtomcat5) provided by the v5.5 packages
and these don't source the setenv.sh in $CATALINA_BASE/bin/, so we have
to do it ourselves in the init script.

This fixes the previous patch which removes the sourcing of setenv by the
init script entirely.