Fix tomcat instance settings on RHEL 7
Problem: RHEL 7 uses systemd unit files to manage the service, which can only
be passed "Environment Files", which are not bash scripts but key=value files
which look like it. This means we cannot use the setenv.sh and setenv-local.sh
files to set env vars. Everything has to be set in the sysconfig/tomcat-xxx
file.